refactor: extract CondaEnvRunner class
